requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
python-dotenv
requests
PyMuPDF
python-docx
pydantic
pytest
httpx
